requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0

plotly>=5.0.0
dash>=2.0.0
